LayoutTests/editing/spelling/document-markers-remain-after-removing-text-expected.txt

PASS internals.hasSpellingMarker(0, 4) is true
PASS successfullyParsed is true

TEST COMPLETE
aaaa bbbb dddd 
